+++ apps/docs/src/content/03-components/structure/list/examples/mobile.tsx
@@ -21,7 +21,7 @@ export default () => {
   const DomainList = typedList<Domain>();
 
   return (
-    <DomainList.List batchSize={2}>
+    <DomainList.List batchSize={2} aria-label="Domains">
       <DomainList.StaticData data={domains} />
       <ActionGroup>
         <Button