+++ apps/remote-dom-demo/src/app/layout.tsx
@@ -60,7 +60,6 @@ export default function Layout(props: PropsWithChildren) {
                 <Link href="/host/simple-form">Simple Form</Link>
                 <Link href="/host/action-form">Action Form</Link>
                 <Link href="/host/react-hook-form">React Hook Form</Link>
-                <Link href="/host/rhf-form">RHF Form</Link>
                 <Link href="/host/suspense">Suspense</Link>
                 <Link href="/host/tunnel">Tunnel</Link>
                 <Link href="/host/ext-bridge">Ext Bridge</Link>