+++ packages/components/src/components/ComboBox/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { ComboBox } from "./ComboBox";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    ComboBox: ViewComponent<typeof ComboBox>;
+  }
+}