+++ packages/components/src/components/List/List.tsx
@@ -1,6 +1,6 @@
-import { ActionGroup } from "@/components/ActionGroup";
 import { DataLoader } from "@/components/List/components/DataLoader";
 import { Header } from "@/components/List/components/Header/Header";
+import headerStyles from "@/components/List/components/Header/Header.module.css";
 import { Items } from "@/components/List/components/Items/Items";
 import { Table } from "@/components/List/components/Table";
 import ListModel from "@/components/List/model/List";
@@ -25,7 +25,6 @@ import {
 } from "@/lib/componentFactory/flowComponent";
 import { type PropsContext, PropsContextProvider } from "@/lib/propsContext";
 import { deepFilterByType, deepFindOfType } from "@/lib/react/deepFindOfType";
-import { deepHas } from "@/lib/react/deepHas";
 import DivView from "@/views/DivView";
 import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
 import type { PropsWithChildren } from "react";
@@ -154,13 +153,15 @@ export const List = flowComponent("List", (props) => {
   const propsContext: PropsContext = {
     ActionGroup: {
       tunnelId: "actions",
+      className: headerStyles.actions,
+      Button: {
+        className: headerStyles.action,
+      },
     },
     ListSummary: {
       tunnelId: "listSummary",
     },
   };
-
-  const hasActionGroup = deepHas(children, ActionGroup);
 
   return (
     <PropsContextProvider props={propsContext}>
@@ -173,7 +174,7 @@ export const List = flowComponent("List", (props) => {
           <DataLoader />
           <DivView className={styles.list} ref={ref}>
             {children}
-            <Header hasActionGroup={hasActionGroup} />
+            <Header />
 
             <DivView className={styles.listWrapper}>
               {listModel.items.entries.length > 0 && (