+++ packages/components/src/components/List/components/Header/components/FilterPicker/index.ts
@@ -1,3 +1,0 @@
-import { FilterPicker } from "./FilterPicker";
-export { FilterPicker } from "./FilterPicker";
-export default FilterPicker;