+++ packages/components/src/components/List/components/Header/components/SortingPickerItem/index.ts
@@ -1,3 +1,0 @@
-import { SortingPickerItem } from "./SortingPickerItem";
-export { SortingPickerItem } from "./SortingPickerItem";
-export default SortingPickerItem;