+++ packages/remote-react-components/src/auto-generated/index.ts
@@ -26,6 +26,7 @@ export * from "./ClearPropsContextContent";
 export * from "./CodeBlock";
 export * from "./Color";
 export * from "./ColumnLayout";
+export * from "./ComboBox";
 export * from "./Content";
 export * from "./ContextMenuContent";
 export * from "./ContextMenuSection";