+++ packages/components/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.196](https://github.com/mittwald/flow/compare/0.2.0-alpha.195...0.2.0-alpha.196) (2025-05-22)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.195](https://github.com/mittwald/flow/compare/0.2.0-alpha.194...0.2.0-alpha.195) (2025-05-22)
+
+### Features
+
+* **Remote:** add remote support to ComboBox ([#1484](https://github.com/mittwald/flow/issues/1484)) ([6c42869](https://github.com/mittwald/flow/commit/6c42869dfa694bd288ce3015628d62e3c2f4af36))
+
+# [0.2.0-alpha.194](https://github.com/mittwald/flow/compare/0.2.0-alpha.193...0.2.0-alpha.194) (2025-05-21)
+
+### Features
+
+* support empty view in ComboBox ([#1476](https://github.com/mittwald/flow/issues/1476)) ([30f6d48](https://github.com/mittwald/flow/commit/30f6d487a2a46741774838f23706acb8c5e0de58))
+
+# [0.2.0-alpha.193](https://github.com/mittwald/flow/compare/0.2.0-alpha.192...0.2.0-alpha.193) (2025-05-21)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.192](https://github.com/mittwald/flow/compare/0.2.0-alpha.191...0.2.0-alpha.192) (2025-05-20)
+
+### Features
+
+* **List:** support context menu and action group in tile ([#1477](https://github.com/mittwald/flow/issues/1477)) ([ffa0c0f](https://github.com/mittwald/flow/commit/ffa0c0ffc293ce08b913597de561c52cb68ad41f))
+
+# [0.2.0-alpha.191](https://github.com/mittwald/flow/compare/0.2.0-alpha.190...0.2.0-alpha.191) (2025-05-20)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.190](https://github.com/mittwald/flow/compare/0.2.0-alpha.189...0.2.0-alpha.190) (2025-05-20)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
 # [0.2.0-alpha.189](https://github.com/mittwald/flow/compare/0.2.0-alpha.188...0.2.0-alpha.189) (2025-05-19)
 
 **Note:** Version bump only for package @mittwald/flow-react-components